feat: Implementación algoritmo Dijkstra Modificado

+++ TransportNetwork.Lib/Graph/Edge.cs
@@ -5,5 +5,5 @@ public class Edge
     public Guid Id { get; set; }
     public Node Source { get; set; }
     public Node Destination { get; set; }
-    public int Cost { get; set; }
+    public float Cost { get; set; }
 }
